+++ src/main/java/com/example/swups/repository/PlansOfStudiesRepository.java
@@ -4,6 +4,8 @@ import com.example.swups.entity.Plansofstudy;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface PlansOfStudiesRepository extends JpaRepository<Plansofstudy, Long> {
+public interface PlansOfStudiesRepository extends JpaRepository<Plansofstudy, Integer> {
 }
