+++ src/main/java/com/example/swups/service/PlansOfStudiesService.java
@@ -18,7 +18,7 @@ public class PlansOfStudiesService {
     public List<Plansofstudy> getPlansOfStudies() {
         return plansOfStudiesRepository.findAll();
     }
-    public Plansofstudy getPlanOfStudiesById(long id) {
+    public Plansofstudy getPlanOfStudiesById(int id) {
         Optional<Plansofstudy> planById = plansOfStudiesRepository.findById(id);
         return planById.orElseThrow(EntityNotFoundException::new);
     }
